% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpkg5-package.R
\docType{package}
\name{testpkg5-package}
\alias{testpkg5}
\alias{testpkg5-package}
\title{testpkg5: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
